HTCONDOR-2908 Fix condor_adstash startd discovery
Instead of using Collector.locate(), add MyAddress to the projection
list when fetching the non-dynamic slot ads the first time, which the
Startd objects can use when being created. This not only fixes the
incompatibility where locate() used to take a Startd slot name instead
of a machine name for its name argument, but is also much, much
faster. We should consider doing this for schedd discovery as well.