HTCONDOR-2582 Make IDTokens that are true capabilities
If a capability IDToken has a scope claim, the client is granted the given
authorization levels, ignoring what's specified in the daemon's ALLOW/DENY
lists. The 'sub' and 'scope' values are still taken from the tokens database
of the daemon.

Capability tokens are now explicitly labeled with a 'cap' claim set to true.

Analysis of the IDToken payload has been moved earlier in theauthentication
protocol, so that the server can inform the client if validation fails
(e.g. capability token not found in the database).