HTCONDOR-2931 Overhaul the UserIdentity class for grid jobs
Rename to make clear it's only for grid universe.
It now contains a pointer to the JobQueueUserRec of the user, which will
simplify adding the OS account that's distinct from the job User.
Pass the GridUserIdentity to GridUniverseLogic, instead of pulling out all
of it's individual components.
Removed dead code for gridmanager-per-job.