HTCONDOR-2931 Add OsUser attribute to job queue records
OsUser is the OS account associated with the given user/job/jobset. On
windows, it's the fully-qualified user account (user@ntdomain). On unix,
it's the username of the account. The value is derived from the User and
NTDomain attributes as before and there should be no change in how the
OS account is selected from the existing code.
This sets the stage to allow the OS account to be independent of the AP's
notion of the job owner (i.e. the User attribute).