HTCONDOR-2975 Don't send collector ad updates via raw UDP command
The UPDATE_COLLECTOR_AD command is moving to DAEMON authorization level,
which means a raw command will almost always fail. We need to do the
normal security sequence.
Use of a raw command was done for the old world ads sent to the central
collector at UW-Madison, which is now long dead.