Add user home page to CE Dashboard. HTCONDOR-2969
Add a landing page designed for sites to use (vs OSG admins) that
allows them to select their own sitename from a list, and store
this site as a default in a browser cookie.

Also, fetch some additionally helpful tidbits from Topology, like the
site 'Name' and if it is a CCStar site or not.