Fix google.charts not defined error in HTCondorView. HTCONDOR-2969
If google charts was not already cached by the browser (e.g. in a Private browser window), and if you are not on a super fast internet connection (e.g. over broadband or 4G), the user may get a google.charts not defined error and have to reload the page.  This happens if Google Charts code does not finish loading before HTCondorView tries to use it.  This patch ensures HTCondorView will wait until Charts is loaded via a Promise.