Improve modal popup on CE Dash overview page. HTCONDOR-2969
On the modal that asks the user if they want this to be their default CE, add a "Stop Asking" button.  Also explain that their answer is stored in cookies, and if they do not want cookies, answer No. (to deal with GDPR requirements... although not certain GDPR applies here... but better error on the safe side)