HTCONDOR-2986 Overhaul handling of User attributes in the job queue
Redo how User is updated in response to a change in UID_DOMAIN. The changes
are now written into the log at startup in a transaction. The UidDomain
attribute in the header ad is used to detect the change in UID_DOMAIN.
Attempts to set Owner or User in proc ads are now diverted to the cluster
ad.
Remove checks after transation commit regarding Owner or User being altered
or set in a proc ad.