Use InstitutionID Names in CE Dash. HTCONDOR-3030
When the CE Dashboard displays the name of the institution, use the new InstitutionID database with standardized institution names instead of the haphazard topology facility names.  These Instituion Names are cached first to disk, so that when using a multi-process server (like gunicorn), only one process needs to fetch and process the database.  Other server processes will read the pickled database from local disk, and then use it to create an in-memory dictionary.