Improve response time of HTCondorView reload. HTCONDOR-3038
Commit ef627d0ab14e2143f081505f401876fbe359cd18 added an option
“reload_page” to the HTCondorViewRanged class.  The idea is to all
graphs on a page to be controlled by one range (day/week/month)
selector, and reload all graphs on the page if the range changes.  This
is used by the CE Dashboard on some pages.  However, the graph
containing the range selector always loaded much slower than all the
other graphs on the page.  This commit fixes that.