HTCONDOR-3041 Add INSTRUCTOR authorization to placementd
The user-login command now takes a 'requester' option, which is the
foreign-id of the user wanting the token. If the requester is different
than the user for whom the token is for, the requester must have the
INSTRUCTOR authorization. The requester is recorded in the placementd's
database for every placement token issued.