Fix delays scheduling jobs leaving cool down. HTCONDOR-3059
Jobs that were in cool down state and are ready to be scheduled may still be ignored by the schedd.  This was because the PrioRecArray is not being marked as 'dirty' in the event a cool down time expires.  Fix this by simply setting a daemon core timer to mark the PrioRecArray as dirty; we can easily and cheaply figure out the time
to use for this timer based upon the scan from our previous build of the PrioRecArray.