HTCONDOR-3060 fix double-free in pipe table
When a daemon exits, we free mallocated memory that contains
the name of the pipe handler and description when when destroy
the pipe table.  Later, in the daemon core destructor, when
we destroy the pid table, the pid entry dtor also tries to
free this memory.  As an imperfect hack, let's just null out
the pointers to this mallocated memory to prevent the double
free.

The PipeTable should contain bona-fide objects with destructors,
and when the PidEntry destructor unregisters any pipes, it shouldn't
complain when it isn't found, but that's work for another day.