Add limit=lastDate mechanism to HTCondorView. HTCONDOR-3076
The limit= keyword now also accepts a keyword "lastDate", which means
to limit the view to all the rows at the bottom of the table that
share the same Date value.  This is useful to limit a view to just
current data to answer questions like "what is happening right now".