Add addRowIfEmpty verb to HTCondorView. HTCONDOR-3076
Add keyword addRowIfEmpty=String, which will add a Row if and only
if the table is empty.  Each numeric column in the new row will
receive a '0', each date column will receive the current date/time,
and each string column will receive the string specified as the
keyword value.  This is useful when you don't want a table/graph
to display 'No Data', but instead show a line at zero, etc.