HTCONDOR-3086 Harden Vault credmon checks for CA file and directory
The CA file and directory are required inputs for urllib3 requests,
but there was a case with a version of the v2 Python bindings where
checking for AUTH_SSL_CLIENT_CAFILE in the config was returning a
blank string instead of saying the param didn't eixst, and this blank
string was being fed in as the CA file to urllib3 requests.

This commit hardens the check for CA file and directory to make sure
that they are not blank strings, and if they are (or are outright
missing from the config), then to fall back on (sensible) defaults.