HTCONDOR-3136 Add option to restrict vault tokens to trusted hosts
If new config knob TRUSTED_VAULT_HOSTS is set, the credd will reject vault
tokens from hosts that aren't in that list. The credd looks at the contents
of the token to determine the hostname.
Also, move the code that determines whether to write a .top or .use file
to the add-only section.