HTCONDOR-3180 Add ping() command to v2 python bindings
This behaves the same as SecMan.ping() in the v1 bindings. It sends a
DC_SEC_QUERY command for the given auth level or cedar command to the
indicated daemon. On success, it returns a ClassAd that's a merger of the
ad returned by the daemon, the cedar security session policy ad, and the
Sock policy ad.