HTCONDOR-3182 Add project names to the placementd
Each user in the placementd can now be associated with a list of projects.
When requesting a placement token, the client can specify a specific
project for the token. The project must be one the user is authorized
to use. Eventually, the project will be baked into the placement token.