HTCONDOR-3147: Deal with history file rotation:
ArchiveMonitor functionality now fully implemented. The ArchiveMonitor uses the 'FileSet' struct and the 'ArchiveChange' struct to track
and communicate changes within the archive to the Librarian. The Librarian uses 'FileSet' structs to locally maintain a FileMap of the
files that are currently in the archive and update the FileMap when files are added, rotated, or deleted. The Librarian finishes reading
the last file it was reading when it went to sleep, and then queues the remaining untracked files for reading/insertion into the database.

Other changes:
    The file 'AggregateSchema.sql' is the schema this implementation uses. It now includes a 'StatusData' table which is meant to
    later be setup to track the health of the status by logging estimated backlog.
    Functions in DBHandler that were purely for testing during development have been moved to DBHandlerTestUtils.cpp.
    I got rid of the JSON library I imported