HTCONDOR-3150 Add garbage collection to database
Added garbage collection policy to database.
Max size is a config with a default set to 2 MB. When database size hits or surpasses the high watermark, which is 97% of capacity,
it calculates the n number of jobs that it needs to delete to get down to the low watermark, which is 85% of capacity.
It then deletes those files. Whether garbage collection is run is also stored as a column in the 'Status' table.