HTCONDOR-3165 Make test work better inside default containers
By default, when running inside a docker container, the hostname in
the container is the sha-1 of the running container, with no dominname.
e.g. 'e34566'.  Some of our tests assume there is a domainname
attached to the hostname, e.g. 'e11234.wisc.edu'.  This fixes
a test that assumes the domainname.