HTCONDOR-3201 Add SecurityContext class to python bindings
This object encapsulates overrides to security configuration that can be
applied to interactions with a specific daemon. At present, the only
option is to provide an in-memory IDToken to be used during authentication.
The ping() function and the Collector class can now accept an optional
SecurityContext. Eventually, all objects that serve as a client to a
daemon should be updated.