Use both code and subcode to decide hold vs vacate. HTCONDOR-2952
The starter passes back errors to the shadow in the form of a
string, a code, and a subcode.  Before this commit, if the
code was greater less than 1000, that told the shadow to place
the job on hold, while code >= 1000 meant no hold (just a VacateCode).
This commit adds examination of the subcode as well, so that
the new rules going forward:
1. Code >= 1000 means vacate instead of hold, OR
2. (NEW WITH THIS COMMIT): Any subcode is less than or equal to -1000 means vacate instead of hold, OR
3. An unspecified error code (code==0) is also vacate instead of hold.