HTCONDOR-2952 Add explicit hold/no-hold suggestion from EP to AP on vacate
When the EP vacates a job, the starter now sends VacateEPSuggestsHold to
indicate whether it believes the job should be held (i.e. is the vacate
due to a defect in the job).
If this new attribute is not sent by the starter, then the AP uses the
old metric for whether to consider holding the job (hold if vacate code
is greater than 0 and less than 1000).