HTCONDOR-3247 Remove retired command GET_PRIORITY_ROLLUP_OLD
This CEDAR command used to share the same integer as DRAIN_JOBS. A new
integer was assigned 12 years ago but the old one is still recognized.
Contrary to the comment in the code, when mapping the integer to the
string name for logging, we get the retired GET_PRIORITY_ROLLUP_OLD
instead of the still-active DRAIN_JOBS.
Let's remove GET_PRIORITY_ROLLUP_OLD fully.