HTCONDOR-3377 Remove ctime check from read_secure_file()
This check will fire if the file is updated via atomic rename during
reading, which is not a problem. It won't catch any legitimate issues
that we care about, so just remove it.
Leave the check of mtime, which can tell us about sloppy file updates
resulting in an incomplete read.