HTCONDOR-3420 Handle short reads when peeking at socket data
When using MSG_PEEK with condor_read(), detect short reads and handle them
appropriately. In condor_read(), don't consider the bytes as consumed when
waiting for more data. In DaemonCommandProtocol, detect a short read and
reject the connection.

Also simplify the parsing of the peeked-at data.

This affects the shared port daemon when it decides whether to transparently
hand a connection off to the collector.