HTCONDOR-3429 Fix file xfer race condition on child exit37
The file transfer reaper fires when the forked child exit.  At this
time, there is certainly a final update for it to read, and there may be
some lingering in-progress updates.  Today, the reaper sets
Info.in_progress = false; at the top of the reaper, but this is
incorrect if it needs to drain some in-progress callbacks before
processing the final update.  If this happens, the starter and shadow
get out of sync with each other after a failed file xfer, and the shadow
goes into reconnect mode, but will never reconnect.  This causes 20
minutes of badput as the shadow hangs around uselessly.

Fix this by only setting `Info.in_progress = false;` when we know we are
calling the file update callback.