Implement knob AUTO_USE_POLICY_PelicanRetryPolicy  HTCONDOR-3475
This knob will default to True for the schedd and false for all other
subsystems.  If True, then the "use policy:PelicanRetryPolicy" will be activated,
which is currently defined to be the same policy as used in the OSPool.
Specifically, job will enter cool down if Pelican fails on upload or download
with specific "retryable" subcodes. This is not a long-term solution.